requirements.txt
# Core Dependencies
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Supabase
supabase==2.0.3

# Cryptography
pycryptodome==3.19.0

# PDF Generation
reportlab==4.0.7
qrcode[pil]==7.4.2
Pillow==10.1.0

# Desktop App (PyQt5)
PyQt5==5.15.10

# Windows-specific (for device detection)
pywin32==306; sys_platform == 'win32'

# Development
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==6.1.0